Add phase service
