+++ src/app/auth/auth.component.ts
@@ -6,6 +6,7 @@ import {HttpErrorResponse} from '@angular/common/http';
 import {ErrorHandlingService} from '../core/services/error-handling.service';
 import {Router} from '@angular/router';
 import {GithubService} from '../core/services/github.service';
+import {PhaseService} from '../core/services/phase.service';
 
 @Component({
   selector: 'app-auth',
@@ -16,15 +17,13 @@ export class AuthComponent implements OnInit, OnDestroy {
   authState: AuthState;
   authStateSubscription: Subscription;
   loginForm: FormGroup;
-  phaseUrl: string;
-  org: string;
-  repo: string;
 
   constructor(private auth: AuthService,
               private github: GithubService,
               private formBuilder: FormBuilder,
               private errorHandlingService: ErrorHandlingService,
-              private router: Router) { }
+              private router: Router,
+              private phaseService: PhaseService) { }
 
   ngOnInit() {
     this.authStateSubscription = this.auth.currentAuthState.subscribe((state) => {
@@ -43,31 +42,13 @@ export class AuthComponent implements OnInit, OnDestroy {
   }
 
   login(form: NgForm) {
-    const full_url = 'full_name';
     if (this.loginForm.invalid) {
       return;
     } else {
       this.auth.startAuthentication(this.loginForm.get('username').value, this.loginForm.get('password').value,
         this.loginForm.get('encodedText').value).subscribe((res) => {
-        let phase;
-        if (res['first']['id'] != null) {
-          this.phaseUrl = '';
-          phase = 'first';
-        } else if (res['second']['id'] != null) {
-          this.phaseUrl = 'phase2';
-          phase = 'second';
-        } else if (res['third']['id'] != null) {
-          this.phaseUrl = 'phase3';
-          phase = 'third';
-        }
-        if (this.phaseUrl == null) {
-          this.errorHandlingService.handleGeneralError('Repo is not ready');
-        } else {
-          this.org = res[phase][full_url].split('/', 2)[0];
-          this.repo = res[phase][full_url].split('/', 2)[1];
-          this.github.updatePhaseDetails(this.repo, this.org);
-          this.router.navigateByUrl(this.phaseUrl);
-        }
+        const phase = this.phaseService.determinePhaseNumber(res);
+        this.router.navigateByUrl(phase);
       }, (error) => {
         if (error instanceof HttpErrorResponse) {
           this.errorHandlingService.handleHttpError(error.error);
