+++ src/app/core/services/auth.service.ts
@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders} from '@angular/common/http';
 import { Router } from '@angular/router';
-import {BehaviorSubject, Observable, of, throwError, from, forkJoin} from 'rxjs';
+import {BehaviorSubject, throwError} from 'rxjs';
 import { NgZone } from '@angular/core';
 import { ElectronService } from './electron.service';
 import {UserService} from './user.service';
+import {PhaseService} from './phase.service';
 import {ErrorHandlingService} from './error-handling.service';
 import {GithubService} from './github.service';
-import {catchError, flatMap, map} from 'rxjs/operators';
+import { flatMap} from 'rxjs/operators';
 import {IssueService} from './issue.service';
 
 export enum AuthState { 'NotAuthenticated', 'AwaitingAuthentication', 'Authenticated' }
@@ -23,50 +24,8 @@ export class AuthService {
               private http: HttpClient,  private errorHandlingService: ErrorHandlingService,
               private githubService: GithubService,
               private userService: UserService,
-              private issueService: IssueService) {
-  }
-
-  parseEncodedPhase(encodedText: String): string[] {
-      const phase = encodedText.split('@', 3);
-      const phaseOneUrl = phase[0].split('=', 2)[1];
-      const phaseTwoUrl = phase[1].split('=', 2)[1];
-      const phaseThreeUrl = phase[2].split('=', 2)[1];
-
-      let separator = phaseOneUrl.lastIndexOf('/');
-      const repoName = phaseOneUrl.substring(separator + 1);
-
-      let separatorOrg = phaseOneUrl.indexOf('.com');
-      const orgName = phaseOneUrl.substring(separatorOrg + 5, separator);
-
-      separator = phaseTwoUrl.lastIndexOf('/');
-      const repoNameSecond = phaseTwoUrl.substring(separator + 1);
-      separatorOrg = phaseTwoUrl.indexOf('.com');
-      const orgNameSecond = phaseTwoUrl.substring(separatorOrg + 5, separator);
-
-    separator = phaseThreeUrl.lastIndexOf('/');
-    const repoNameThird = phaseThreeUrl.substring(separator + 1);
-    separatorOrg = phaseThreeUrl.indexOf('.com');
-    const orgNameThird = phaseThreeUrl.substring(separatorOrg + 5, separator);
-
-      return new Array(repoName, orgName, repoNameSecond, orgNameSecond, repoNameThird, orgNameThird);
-  }
-
-  checkIfReposAccessible(array: any): any {
-
-    const url1 = 'https://api.github.com/repos/' + array[1] + '/' + array[0];
-    const url2 = 'https://api.github.com/repos/' + array[3] + '/' + array[2];
-    const url3 = 'https://api.github.com/repos/' + array[5] + '/' + array[4];
-
-    const value = forkJoin(
-      this.http.get(url1).pipe(map((res) => res), catchError(e => of('Oops!'))),
-      this.http.get(url2).pipe(map((res) => res), catchError(e => of('Oops!'))),
-      this.http.get(url3).pipe(map((res) => res), catchError(e => of('Oops!'))),
-    ).pipe(
-      map(([first, second, third]) => {
-        return {first, second, third};
-      })
-    );
-    return value;
+              private issueService: IssueService,
+              private phaseService: PhaseService) {
   }
 
   startAuthentication(username: String, password: String, encodedText: String) {
@@ -80,9 +39,9 @@ export class AuthService {
       }),
       flatMap((user) => {
         if (user) {
-          const array = this.parseEncodedPhase(encodedText);
+          const array = this.phaseService.parseEncodedPhase(encodedText);
           this.changeAuthState(AuthState.Authenticated);
-          return (this.checkIfReposAccessible(array));
+          return (this.phaseService.checkIfReposAccessible(array));
         } else {
           return throwError('Unauthorized user.');
         }
